Completed Visual Design of Website
Really like it. Now have to begin to refocus in on the javascript. Specifically getting the play the game button to start the play game function. Then getting the #’s of wins to update per user and computer win total. Then getting the boxes of top and bottom to change text based on situation.