Working through the JS more
Got some of the function working. Stuck on getting the player’s choice. Working through the JS makes me more aware of how I should change the CSS to better design the website, and how I really need to affirm my understanding of CSS types so as to better utilize the organization in CSS.